Creating and Styling NftCard Component #33
The task involves creating the NftCard component, which will be used to represent NFT items in the form of cards. The goal is to establish the basic structure of the card and refine its content. The component should include an image representing the NFT, the NFT's name, the publisher, and the current bid.

Co-Authored-By: Faru Nuri Sönmez <info@farunurisonmez.com>