Added zod package #51
Zod is added to facilitate the validation and definition of environment variables in the project. With this change, the Zod package is included to streamline the validation process for environment variables, ensuring that values in the .env file are in the correct format and complete, utilizing Zod's schema validation features.

Close #51

Co-Authored-By: Faru Nuri Sönmez <info@farunurisonmez.com>